requirements.txt
pillow>=9.0.0
pytesseract>=0.3.10
plyer>=2.1.0
